TypedArray prototype set should go down the fast path when using non clamped integer types of the same byte size https://bugs.webkit.org/show_bug.cgi?id=233905
Reviewed by Keith Miller.

JSTests:

* microbenchmarks/typed-array-prototype-set.js: Added.
* microbenchmarks/typed-array-prototype-set-order.js: Added.
* stress/typed-array-prototype-set.js: Added.

Source/JavaScriptCore:

We can use memmove in this scenario because the bitpattern of the
data between the signed and unsigned values will be the same.

This patch also fixes a bug where we were looking at the wrong
pointer when determining to do a forward or backwards loop in
our memmove. We were looking at the vector instead of vector+offset.

* runtime/JSGenericTypedArrayViewInlines.h:
(JSC::JSGenericTypedArrayView<Adaptor>::set):

LayoutTests:

* js/script-tests/typedarray-set-overlapping-elements-of-same-size.js:
* js/typedarray-set-overlapping-elements-of-same-size-expected.txt:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@286639 268f45cc-cd09-0410-ab3c-d52691b4dbfc