Same-site lax cookies not sent by fetch event handler after page reload https://bugs.webkit.org/show_bug.cgi?id=226386 <rdar://problem/78878853>
Reviewed by Chris Dumez.

Source/WebCore:

When a service worker fetches a navigation request exposed from the fetch event, we need to keep isTopSite intact as
the service worker does not really have the information of which frame is actually loaded and whether it is a main frame or not.

Tests: http/wpt/service-workers/cross-site-navigation-same-cookie-lax.https.html
       http/wpt/service-workers/same-cookie-lax.https.html

* loader/cache/CachedResource.cpp:

Source/WebKit:

StorageBlockingPolicy handling is not covered by generated code so explicit update StorageBlockingPolicy setting from preference store.
This impacts the computation of cookie/cache partitioning.

* WebProcess/Storage/WebSWContextManagerConnection.cpp:

LayoutTests:

* http/wpt/service-workers/cross-site-navigation-same-cookie-lax.https-expected.txt: Added.
* http/wpt/service-workers/cross-site-navigation-same-cookie-lax.https.html: Added.
* http/wpt/service-workers/resources/get-cookie.py: Added.
* http/wpt/service-workers/resources/get-document-cookie.py: Added.
* http/wpt/service-workers/resources/set-cookie-lax.py: Added.
* http/wpt/service-workers/same-cookie-lax-worker.js: Added.
* http/wpt/service-workers/same-cookie-lax.https-expected.txt: Added.
* http/wpt/service-workers/same-cookie-lax.https.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@286656 268f45cc-cd09-0410-ab3c-d52691b4dbfc