Teach webpushd to launch an app in response to an injected message. https://bugs.webkit.org/show_bug.cgi?id=234122
Reviewed by Alex Christensen.

Source/WebCore/PAL:

* pal/spi/cocoa/LaunchServicesSPI.h:

Source/WebKit:

To by covered by future API test.

When a message comes in, webpushd knows which code signing identifier it was targetted for.
It launches that app in a special mode so the app can then "take it from there"

* webpushd/WebPushDaemon.h:
* webpushd/WebPushDaemon.mm:
(WebPushD::Daemon::injectPushMessageForTesting):
(WebPushD::Daemon::notifyClientPushMessageIsAvailable):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@286878 268f45cc-cd09-0410-ab3c-d52691b4dbfc