[Cocoa] OT-SVG glyphs don't draw into canvases (because of the GPU process) https://bugs.webkit.org/show_bug.cgi?id=234171 <rdar://problem/70166552>
Reviewed by Devin Rousso.

Source/WebCore:

Drawing OT-SVG glyphs into canvas was intentionally disabled in https://trac.webkit.org/changeset/269211/webkit.
This patch enables it again. Rather than doing anything complicated like supporting all of SVG in DrawGlyphsRecorder,
we can simply support this by drawing the glyphs into a ImageBuffer and sending the ImageBuffer to the GPU process.

For text, it's pretty important that the pixel grid of the ImageBuffer matches the pixel grid of the destination,
rather than being offset by half a pixel or something. This patch adds a new creation function to ImageBuffer which
accepts a FloatRect (instead of the previous FloatSize which it used to accept). The FloatRect is necessary because
inflating the geometry has to happen on both the left and the right if we want the pixel grids to match.

Test: fast/text/otsvg-canvas.html

* platform/graphics/DrawGlyphsRecorder.h:
* platform/graphics/ImageBuffer.cpp:
(WebCore::ImageBuffer::createCompatibleBuffer):
(WebCore::ImageBuffer::compatibleBufferInfo):
* platform/graphics/ImageBuffer.h:
* platform/graphics/coretext/DrawGlyphsRecorderCoreText.cpp:
(WebCore::DrawGlyphsRecorder::drawOTSVGRun):
(WebCore::DrawGlyphsRecorder::drawNonOTSVGRun):
(WebCore::DrawGlyphsRecorder::drawBySplittingIntoOTSVGAndNonOTSVGRuns):
(WebCore::DrawGlyphsRecorder::drawGlyphs):
(WebCore::filterOutOTSVGGlyphs): Deleted.

LayoutTests:

* fast/text/otsvg-canvas-expected.html: Added.
* fast/text/otsvg-canvas.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@286889 268f45cc-cd09-0410-ab3c-d52691b4dbfc