A FormData constructed in the form's submit event listener shouldn't include the submitter https://bugs.webkit.org/show_bug.cgi?id=234069
Patch by Andreu Botella <andreu@andreubotella.com> on 2021-12-13
Reviewed by Chris Dumez.

LayoutTests/imported/w3c:

* web-platform-tests/html/semantics/forms/form-submission-0/constructing-form-data-set-expected.txt:

Source/WebCore:

In the HTML spec, the "construct the entry list" algorithm takes an optional `submitter`
argument which it uses to determine whether a button should be included in the entry list.
The FormData constructor calls it with no optional arguments, so `new FormData(form)` should
never contain any buttons.

In WebKit, however, if a `FormData` object is constructed in the form's `submit` event
listener, it will contain the submitter button, if the form submission wasn't implicit. This
is because whether a form control is the submitter is tracked on the control, and activating
it sets that state for the duration of the form submission algorithm.

However, the form submission algorithm also sets a submitter's state, to deal with implicit
submissions. This is set only for the duration of the "construct the entry list" call, and
so there is no need to set the state when activating the button.

Tests: imported/w3c/web-platform-tests/html/semantics/forms/form-submission-0/constructing-form-data-set.html

* html/HTMLButtonElement.cpp:
(WebCore::HTMLButtonElement::defaultEventHandler):
* html/ImageInputType.cpp:
(WebCore::ImageInputType::handleDOMActivateEvent):
* html/SubmitInputType.cpp:
(WebCore::SubmitInputType::handleDOMActivateEvent):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@286988 268f45cc-cd09-0410-ab3c-d52691b4dbfc