Web Inspector: After r286657, anonymous layers are providing an empty string for their grouping text https://bugs.webkit.org/show_bug.cgi?id=234012
Reviewed by Devin Rousso.

Covered by existing test: LayoutTests/inspector/css/getMatchedStylesForNodeLayerGrouping.html

For anonymous layers, we should not provide the `text` property to the frontend, as there is technically no
text, not an empty string, provided by the author. This is effectively a minor optimization, as the frontend
does correctly handle an empty string the same way it handles no string.

* inspector/InspectorStyleSheet.cpp:
(WebCore::buildArrayForGroupings):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287046 268f45cc-cd09-0410-ab3c-d52691b4dbfc