Refactor WidthIterator::applyCSSVisibilityRules() to be a little more elegant https://bugs.webkit.org/show_bug.cgi?id=234428
Reviewed by Alan Bujtas.

This adds a few lambda operations in this function:
- adjustForSyntheticBold()
- clobberGlyph()
- clobberAdvance()
- deleteGlyph()
And then updates the rest of the function to use those operations. I think this makes
the body of the function a little easier to understand, because the code uses a higher
level of abstraction. It shoudln't be any slower, though, because the lambdas can all
be inlined to the same code that was there before.

No new tests because there is no behavior change.

* platform/graphics/WidthIterator.cpp:
(WebCore::WidthIterator::applyCSSVisibilityRules):
(WebCore::WidthIterator::adjustForSyntheticBold): Deleted.
* platform/graphics/WidthIterator.h:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287208 268f45cc-cd09-0410-ab3c-d52691b4dbfc