Keyboard shortcut to scroll to top when already at the top of the page moves to the bottom https://bugs.webkit.org/show_bug.cgi?id=234483 <rdar://86628260>
Reviewed by Dean Jackson.
Source/WebCore:

If the page was scrolled to the top and an "up" keyboard scroll happened,
ScrollAnimator::singleAxisScroll() would trigger an unclamped scroll with a negative delta,
which fed into ScrollAnimationSmooth::startAnimatedScrollToDestination() and would result in
an animation with a zero duration, which resulted in NaNs in animateScroll().

Fix by doing clamping in ScrollAnimator::singleAxisScroll() and protecting against
animations with zero delay in ScrollAnimationSmooth.

Test: fast/scrolling/keyboard-scrolling-home.html

* platform/ScrollAnimationSmooth.cpp:
(WebCore::ScrollAnimationSmooth::startAnimatedScrollToDestination):
(WebCore::ScrollAnimationSmooth::retargetActiveAnimation):
* platform/ScrollAnimator.cpp:
(WebCore::ScrollAnimator::singleAxisScroll):

LayoutTests:

* fast/scrolling/keyboard-scrolling-home-expected.txt: Added.
* fast/scrolling/keyboard-scrolling-home.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287241 268f45cc-cd09-0410-ab3c-d52691b4dbfc